Move add-clinic-rules to the `validationRules` folder
